metaai-sdk/requirements.txt
httpx>=0.25.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pydantic[v2]>=2.7.0
python-dotenv>=1.0.0